requirements.txt
alembic==1.17.2
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Assets==2.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
setuptools==80.9.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
uWSGI==2.0.31
webassets==3.0.0
Werkzeug==3.1.4
WTForms==3.2.1
zope.dottedname==7.0
